Paylaod validation
Simple payload validation that stores 1 error in the redis cache, which
can then be obtained by the tennant via the status endpoint.

This is an initial setup and all naming and error position is up for
grabs

At the moment this fails fast and provides 1 error. This would be used
for errors like:
- No file
- Mime types
- File to large